+++ src/backend/VoltStream.Application/Features/Supplies/DTOs/SupplyDTO.cs
@@ -7,7 +7,7 @@ public record SupplyDto
 {
     public long Id { get; set; }
     public DateTimeOffset Date { get; set; }
-    public decimal RollCount { get; set; }
+    public int RollCount { get; set; }
     public decimal LengthPerRoll { get; set; }
     public decimal TotalLength { get; set; }
     public decimal UnitPrice { get; set; }
